refactor GET and POST /meals
currently failing because Goals must be refactored first
