+++ lib/routes/api/v1/meals.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+const mealsController = require('../../../controllers/meals_controller');
+
+router.get('/', mealsController.index);
+
+router.post('/', mealsController.create);
+
+module.exports = router
