+++ server.js
@@ -2,6 +2,7 @@ const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
 const foods = require('./lib/routes/api/v1/foods')
+const meals = require('./lib/routes/api/v1/meals')
 
 const environment = process.env.NODE_ENV || 'development';
 const configuration = require('./knexfile')[environment];
@@ -34,41 +35,7 @@ app.use('/api/v1/foods/:id', foods)
 
 // -------Meals-------
 
-app.get('/api/v1/meals', (request, response) => {
-  database('meals')
-  .leftJoin('meal_foods', 'meal_foods.meal_id', '=', 'meals.id')
-  .leftJoin('foods', 'meal_foods.food_id', '=', 'foods.id')
-  .select( 'meals.id AS meal_id', 'meals.name AS meal_name', 'meals.date AS meal_date', 'meals.calorie_goal AS meal_goal', 'foods.id AS food_id', 'foods.name AS food_name', 'foods.calories AS food_calories')
-    .then((meals_data) => {
-      let meals = [];
-      meals_data.forEach( data => {
-        if ( meals.find( meal => meal.id == data.meal_id)) {
-          let existing_meal = meals.find( meal => meal.id == data.meal_id);
-          existing_meal.foods.push({
-            id: data.food_id,
-            name: data.food_name,
-            calories: data.food_calories
-          })
-        } else {
-          meals.push({
-            id: data.meal_id,
-            name: data.meal_name,
-            date: data.meal_date,
-            calorie_goal: data.meal_goal,
-            foods: [{
-              id: data.food_id,
-              name: data.food_name,
-              calories: data.food_calories
-            }]
-          })
-        }
-      })
-      response.status(200).json(meals);
-    })
-    .catch((error) => {
-      response.status(500).json({ error });
-    });
-});
+app.use('/api/v1/meals', meals)
 
 app.get('/api/v1/meals/:meal_id/foods', (request, response) => {
   database('meals')
@@ -113,52 +80,52 @@ app.get('/api/v1/meals/:meal_id/foods', (request, response) => {
     });
 })
 
-app.post('/api/v1/meals', (request, response) => {
-  const meal_data = request.body.meal;
-  const meal_name = request.body.meal.name;
-  const meal_date = request.body.meal.date;
-
-  for (let requiredParameter of ['name', 'date']) {
-    if (!meal_data[requiredParameter]) {
-      return response
-        .status(422)
-        .send({ error: `Expected format: { meal: { name: <String>, date: <Date> } }. You're missing a "${requiredParameter}" property.` });
-    }
-  }
-  database('meals')
-    .where({name: meal_name, date: meal_date})
-    .select()
-    .then( meals => {
-      if (meals.length) {
-        return response
-          .status(422)
-          .send({ error: `Meal already exists for that Date.` });
-      } else {
-        database('goals')
-        .where({name: meal_name})
-        .select()
-        .then( goals => {
-          let calorie_goal = goals[0].calories;
-          database('meals').insert({name: meal_name, date: meal_date, calorie_goal: calorie_goal}, 'id')
-          .then(meal_id => {
-            database('meals').where({id: meal_id[0]}).select()
-            .then(meal => {
-              response.status(201).json(meal[0]);
-            })
-          })
-          .catch( error => {
-            response.status(500).json({ error });
-          })
-        })
-        .catch( error => {
-          response.status(500).json({ error });
-        })
-      }
-    })
-    .catch( error => {
-      response.status(500).json({ error });
-    })
-})
+// app.post('/api/v1/meals', (request, response) => {
+//   const meal_data = request.body.meal;
+//   const meal_name = request.body.meal.name;
+//   const meal_date = request.body.meal.date;
+//
+//   for (let requiredParameter of ['name', 'date']) {
+//     if (!meal_data[requiredParameter]) {
+//       return response
+//         .status(422)
+//         .send({ error: `Expected format: { meal: { name: <String>, date: <Date> } }. You're missing a "${requiredParameter}" property.` });
+//     }
+//   }
+//   database('meals')
+//     .where({name: meal_name, date: meal_date})
+//     .select()
+//     .then( meals => {
+//       if (meals.length) {
+//         return response
+//           .status(422)
+//           .send({ error: `Meal already exists for that Date.` });
+//       } else {
+//         database('goals')
+//         .where({name: meal_name})
+//         .select()
+//         .then( goals => {
+//           let calorie_goal = goals[0].calories;
+//           database('meals').insert({name: meal_name, date: meal_date, calorie_goal: calorie_goal}, 'id')
+//           .then(meal_id => {
+//             database('meals').where({id: meal_id[0]}).select()
+//             .then(meal => {
+//               response.status(201).json(meal[0]);
+//             })
+//           })
+//           .catch( error => {
+//             response.status(500).json({ error });
+//           })
+//         })
+//         .catch( error => {
+//           response.status(500).json({ error });
+//         })
+//       }
+//     })
+//     .catch( error => {
+//       response.status(500).json({ error });
+//     })
+// })
 
 app.post('/api/v1/meals/:meal_id/foods/:food_id', (request, response) => {
   let meal_id = request.params.meal_id;
